requirements.txt
aiofiles
aiohttp
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
httpx
motor>=3.6.0
pymongo[srv,tls]>=4.9.1
numpy
opencv-python
pillow
psutil
py-tgcalls
pytgcrypto
#pyrogram
kurigram
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
unidecode
yt-dlp
youtube-search
py-yt-search
telegraph